add graceful shutdown on sigterm for client
