+++ examples/example_task.json
@@ -1,6 +1,6 @@
 {
   "url": [
-    "cnn.com"
+    "kumarde.com"
   ],
   "browser_settings": {
     "browser_binary": "",
